cleanups WIP
- Don't use symbols that start with "_" followed by a capital letter
  (e.g. _SYMBOLTABLE_H). These are reserved symbols by the C++ standard.
  Same for all symbols that contain "__" (double underscore).
- Bug fixes in Convert.cpp
   code could crash on (incomplete) strings like '&' or 'H'.
- Various small changes in DockManager

git-svn-id: file:///openmsx-svn/code/openmsx-debugger/trunk@9339 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/VramBitMappedView.cpp
@@ -237,7 +237,7 @@ void VramBitMappedView::decodeSCR5()
 void VramBitMappedView::paintEvent(QPaintEvent* event)
 {
 	QRect srcRect(0, 0, 512, 2 * lines);
-	QRect dstRect(0, 0, 512 * zoomFactor, 2 * lines * zoomFactor);
+	QRect dstRect(0, 0, int(512 * zoomFactor), int(2 * lines * zoomFactor));
 	QPainter qp(this);
 	//qp.drawImage(rect(),image,srcRect);
 	qp.drawPixmap(dstRect, piximage, srcRect);
